refactor(library/URI): moves "/UniformResourceIdentifier" to "/URI"
