+++ src/features/TextToImage/Pipeline/Output/Image.ts
@@ -1,6 +1,6 @@
 import type {
   URI_Image,
-} from '@/library/UniformResourceIdentifier';
+} from '@/library/URI';
 
 interface TextToImage_Pipeline_Output_Image {
   uri: URI_Image;
