+++ src/library/URI/Data/EncodingIdentifier/Any.ts
