+++ src/library/URI/Data/EncodingIdentifier/all.ts
