+++ src/library/URI/Data/EncodingIdentifier/exports_objectOriented.ts
