+++ src/library/URI/Data/EncodingIdentifier/index.ts
