+++ src/library/URI/Data/EncodingIdentifier/namespace.ts
