+++ src/library/URI/Data/definition.ts
