+++ src/library/URI/Data/definitionAugmentation.ts
