+++ src/library/URI/Data/definitionWithAugmentation.ts
