+++ src/library/URI/Data/exports_objectOriented.ts
