+++ src/library/URI/Data/index.ts
