+++ src/library/URI/Image/Format/Any.ts
