+++ src/library/URI/Image/Format/all.ts
