+++ src/library/URI/Image/Format/exports_objectOriented.ts
