+++ src/library/URI/Image/Format/index.ts
