+++ src/library/URI/Image/Format/namespace.ts
