+++ src/library/URI/Image/Format/namespaceMembers.ts
