+++ src/library/URI/Image/definition.ts
