+++ src/library/URI/Image/exports_objectOriented.ts
