+++ src/library/URI/Image/index.ts
